pyproject.toml
[project]
name = "agent-memory-examples"
version = "0.1.0"
description = "Amazon Bedrock Agent Core Memory Management - Unified Streamlit Application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bedrock-agentcore>=1.0.7",
    "boto3>=1.41.5",
    "strands-agents>=1.18.0",
    "streamlit>=1.31.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
]
